fix: 관심분석가 하나만 나오는 문제 수정

+++ backend/stockolm/src/main/java/com/example/stockolm/domain/follow/dto/response/FollowAnalystResponse.java
@@ -8,12 +8,13 @@ public class FollowAnalystResponse {
     private String userName;
     private String userNickName;
     private String userImagePath;
-    private float reliability;
-    private float accuracy;
-    private int totalAnalystRanking;
+    private Double reliability;
+    private Double accuracy;
+    private Integer totalAnalystRanking;
 
     @Builder
-    public FollowAnalystResponse(String userName, String userNickName, String userImagePath, float reliability, float accuracy, int totalAnalystRanking) {
+    public FollowAnalystResponse(String userName, String userNickName, String userImagePath,
+                                 Double reliability, Double accuracy, Integer totalAnalystRanking) {
         this.userName = userName;
         this.userNickName = userNickName;
         this.userImagePath = userImagePath;
